+++ NEWS.md
@@ -3,3 +3,5 @@
 * Fixed a bug in ipums_shape_*_join functions when using integer ID columns. (#16)
 
 * Allow for unzipped folders because Safari on macOS unzips folders by default (#17)
+
+* lbl_relabel behavior is improved so that labels aren't assigned sequentially (#21)